Refactor erf_cody.rs and implement complementary error function
This update refactors erf_cody.rs module, streamlining the number representations for improved readability. The major change is the implementation of the complementary error function, or erfc. This function is a key element in statistical analysis, especially in computing the normal distribution. The implementation follows the algorithm outlined in W.J. Cody's 1969 paper, "Rational Chebyshev Approximations for the Error Function".